// frontend/src/types.ts
export interface AuthUser {
  id: string;
  name: string;
  email: string;
  phone?: string;
  dob?: string;
  role: 'patient' | 'caregiver' | 'clinician' | 'pharmacy';
  insuranceName?: string;
  memberId?: string;
  isSeniorEligible?: boolean;
  avatarInitials: string;
}